mark multiple-open func tests as TODO on Linux
Three tests in UpdatePlaceholderTests and two in
BasicFileSystemTests currently fail due to contention with
libprojfs for an exclusive flock(2).

The C# .NET file locking and sharing implementation on
Linux conflicts with the flock(2) locks taken by
libprojfs when testing the projection status of a file.

Since projection status is checked on every file syscall,
e.g., on open(2), attempts to re-open a file fail because
.NET creates its own lock after the first fd is opened,
and then during the second open(2) libprojfs wants its
own temporary but exclusive lock, which fails and causes
open() to fail, which is then reported by the .NET library
as: "The process cannot access the file '...' because
it is being used by another process."

We can remove this TODO category once libprojfs uses an
internal locking mechanism instead of flock(2).